Extract header and card components
