+++ client/src/components/card/card.jsx
@@ -1,0 +1,3 @@
+export default function Card({ children }) {
+  return <div className="card">{children}</div>;
+}
