+++ client/src/components/header/header.jsx
@@ -1,0 +1,11 @@
+export default function AppHeader() {
+  return (
+    <header>
+      <div className="logo" aria-hidden="true" />
+      <div>
+        <p style={{ margin: 0, color: '#475569' }}>FullStack Pilot</p>
+        <h1>App Manager</h1>
+      </div>
+    </header>
+  );
+}
